Fixes bug which made pepper not work
Fun fact: When crypt() generates hashed password, it uses the second
argument (a $salt) to decide which algorithm to use. $6$ denotes
SHA-512.

However, in a string, using an unescaped $ means to insert a variable
until next breaking character. This means "$6$dankmemes" would be the
characters $6 (a number can't start a character so this is "escaped")
followed by the variable $dankmemes. Yes, this is pretty damn stupid.

This should fix it by adding the SHA-512 characters in the config file,
not when generating it. I didn't see this earlier, since my test
generation started the pepper with a number i.e. properly escaped the
$...