Moves election session closing and opening to actions/adminpagehandler.php
Only the admin (priv == 0) should be able to open and close sessions.
However, for some reason, this is handled in the page handler for
Election Admins (who only handles rounds). This moves code to the right
place, changes POST address in the index/admin.

Login was removed from adminpagehandler, since this was unused code
anyway.